requirements.txt
astroid==2.1.0
certifi==2018.11.29
chardet==3.0.4
clarifai==2.5.2
colorama==0.4.1
configparser==3.7.1
future==0.17.1
googleapis-common-protos==1.5.6
grpcio==1.18.0
idna==2.8
isort==4.3.4
jsonschema==2.6.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.1
opencv-python==3.4.5.20
protobuf==3.6.1
pylint==2.2.2
pyserial==3.4
requests==2.21.0
six==1.12.0
urllib3==1.24.1
wrapt==1.11.1